Eliminate potential crash (uninitialized memory)
The C64 KERNAL recognizes a BRK with LDA $0104,X in it's IRQ routine. If the
stack pointer is low enough it will read from $200-$203 instead of the stack.
If there's the wrong sort of garbage there (bit 4 set) it will interrupt the
test.

Therefore this patch cleans this area to avoid this. $200-$203 is the BASIC
input buffer and normally we are lucky that it's content doesn't make any
trouble. But we should not depend on luck ;)