Gtk3: Refactor the main sound settings UI
Make the various widgets related to main sound settings use the current widgets. Move uisound.* to widgets since it's a child of uisettings.c and not called from the menu and therefor I think it doesn't belong in the root of src/arch/gtk3 but rather in the widgets/ subdir.