Gtk3: Add a resource-browser widget
This widget is basically a combination of a resource_entry_full widget and a browse button and an optional label. A combination I've had to implement a lot, so it makes sense to implement a widget that handles this with a lot fewer calls.
Currently it isn't complete, but at least it will revert a resource when setting that resource fails. Functions like reset() and update() are required but not yet implemented.