Gtk3: VSID: Remember last used dir
Works a little flaky with Gtk's "last opened files" list, which seems to be the default 'dir' for Gtk file choosers. But after selecting on or more files from that 'dir', the last used dir is set to the basedir of the first/only selected file, so stuff works after that.