Remove extra fflush(stdout) calls, not required for linenoise-ng on Windows
Using cmd.exe, linenoise-ng works the same way as on Linux, and no extra fflush() calls are required. I tested c151.exe in the MSYS2 shell, and that seems to confuse the hell out of linenoise-ng.
Would be nice to have linenoise-ng recognize the MSYS2 shell, so debugging c1541 can be done from MSYS, a TODO for now though.