Gtk3: display fliplist activity in the statusbar
See bug #1129. Gtk3's fliplist now shows on the statusbar what it's doing. Unfortunately the statusbar's area for messages is a little small, so not useful info is actually displayed.
Perhaps the GtkLabel can be told to use multiline, or can be added as a third row in the statusbar.